[Intl] Renamed BundleCompiler to GenrbBundleCompiler
