+++ src/Symfony/Component/Intl/ResourceBundle/Compiler/BundleCompiler.php
@@ -11,61 +11,14 @@
 
 namespace Symfony\Component\Intl\ResourceBundle\Compiler;
 
-use Symfony\Component\Intl\Exception\RuntimeException;
-
 /**
- * Compiles .txt resource bundles to binary .res files.
+ * Alias of {@link GenrbBundleCompiler}.
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
+ *
+ * @deprecated Deprecated since version 2.4, to be removed in Symfony 3.0. Use
+ *             the equivalent {@link GenrbBundleCompiler} instead.
  */
-class BundleCompiler implements BundleCompilerInterface
+class BundleCompiler extends GenrbBundleCompiler
 {
-    /**
-     * @var string The path to the "genrb" executable.
-     */
-    private $genrb;
-
-    /**
-     * Creates a new compiler based on the "genrb" executable.
-     *
-     * @param string $genrb   Optional. The path to the "genrb" executable.
-     * @param string $envVars Optional. Environment variables to be loaded when
-     *                        running "genrb".
-     *
-     * @throws RuntimeException If the "genrb" cannot be found.
-     */
-    public function __construct($genrb = 'genrb', $envVars = '')
-    {
-        exec('which ' . $genrb, $output, $status);
-
-        if (0 !== $status) {
-            throw new RuntimeException(sprintf(
-                'The command "%s" is not installed',
-                $genrb
-            ));
-        }
-
-        $this->genrb = ($envVars ? $envVars . ' ' : '') . $genrb;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function compile($sourcePath, $targetDir)
-    {
-        if (is_dir($sourcePath)) {
-            $sourcePath .= '/*.txt';
-        }
-
-        exec($this->genrb.' --quiet -e UTF-8 -d '.$targetDir.' '.$sourcePath, $output, $status);
-
-        if ($status !== 0) {
-            throw new RuntimeException(sprintf(
-                'genrb failed with status %d while compiling %s to %s.',
-                $status,
-                $sourcePath,
-                $targetDir
-            ));
-        }
-    }
 }
